File upload loading working in plot_streamlit.py
Added a separate parsing/reading of uploaded file to the routine.  Command line given files use file_io.load.  If uploaded, this code uses a separate load routine that pulls from the uploaded_file in the BytesIO buffer.  Works for json, cbor, and yaml.  Need to test/get it working for XLSX